perf: load chart and map views only when needed
